pyproject.toml
[project]
name = "tomoji"
version = "1.0.0"
description = "Generate emojis from your face and use them on your website. "
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "mediapipe>=0.10.0",
    "opencv-python-headless>=4.8.0",
    "Pillow>=10.0.0",
    "fonttools[woff]>=4.44.0",
    "brotli>=1.1.0",
    "python-multipart>=0.0.6",
    "slowapi>=0.1.9",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
